+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## 2.9.0
 * Initialize array defaults for primitive types as value array
+* new parameter `useConstants: true`. When set to true, values with one single enum and a matching default value will be generated as constants instead of enums. The default is false, to keep backwards compatibility. 
 
 ## 2.8.1
 * Fix "An expression tree cannot contain pattern-matching 'is' expression"